% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dados.R
\name{consolidar_base}
\alias{consolidar_base}
\title{Consolidar dados API e CN}
\usage{
consolidar_base(api, acs)
}
\arguments{
\item{api}{Tabela de dados das apis}

\item{acs}{Tabela com dados do Cade em Numeros}
}
\value{
A tabela usada para modelagem e previsoes
}
\description{
Consolidar dados API e CN
}
