% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estacionariedade.R
\name{estacionariedade}
\alias{estacionariedade}
\title{Teste de estacionaridade das variaveis}
\usage{
estacionariedade(df)
}
\arguments{
\item{df}{Data frame com variaveis para testar}
}
\value{
Data frame com resultado dos testes
}
\description{
Teste de estacionaridade das variaveis
}
