% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{teste_12_lags}
\alias{teste_12_lags}
\title{Calcula os testes para a variável x em todos 12 lags (sem lag também)}
\usage{
teste_12_lags(x)
}
\arguments{
\item{x}{}
}
\value{

}
\description{
Calcula os testes para a variável x em todos 12 lags (sem lag também)
}
