% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ur.df.modified.R
\name{ur.df.modified}
\alias{ur.df.modified}
\title{ur.df modificado}
\usage{
ur.df.modified(
  y,
  type = c("none", "drift", "trend"),
  lags = 0,
  selectlags = c("Fixed", "AIC", "BIC")
)
}
\arguments{
\item{y}{vetor}

\item{type}{tipo}

\item{lags}{lag}

\item{selectlags}{tipo}
}
\value{
Uma lista
}
\description{
ur.df modificado
}
